+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 # Change log for Microsoft365DSC
 
-# UNRELEASED
+# 1.25.319.1
 
 * AADServicePrincipal
-  * Removed logic that would translate AppId as a disply name from the
-    Set-TargetResource function.
+  * Streamlined logic to translate AppId from a display name.
 * IntuneDeviceConfigurationPolicyAndroidDeviceOwner
   * Fixed an issue where empty complex objects result in a broken export.
     FIXES [#5863](https://github.com/microsoft/Microsoft365DSC/issues/5863)
+* SCSensitivityLabel
+  * Fix an issue where the Priority property was always getting updated for an
+    existing label, even if it didn't get updated.
 * SHSpaceGroup
   * Initial release.
 * SHSpaceUser
@@ -23,6 +25,7 @@
     FIXES [#3158](https://github.com/microsoft/Microsoft365DSC/issues/3158)
 * DEPENDENCIES
   * Updated MicrosoftTeams to version 6.9.0.
+  * Updated MSCloudLoginAssistant to version 1.1.42.
 
 # Unreleased